backend/requirements.txt
aistudio_sdk==0.3.5
annotated-types==0.7.0
anyio==4.10.0
bce-python-sdk==0.9.42
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
fastapi==0.112.0
filelock==3.19.1
fsspec==2025.7.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
loguru==0.7.2
modelscope==1.29.1
networkx==3.5
numpy==2.3.2
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.0
pandas==2.3.2
phonenumbers==8.13.39
pillow==11.3.0
pluggy==1.6.0
prettytable==3.16.0
protobuf==6.32.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
pypdfium2==4.30.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.5
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
